Connect all the classes to database
Polished indentation a little bit, so the midification may seems like I
modified very line of the code. For the mockDao and mockImplementation,
I commented some lines to avoid reporing error. I tested them using
Swagger and they works fine.

+++ src/main/java/team/awesome/pet/controller/ShelterController.java
@@ -12,37 +12,40 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import team.awesome.pet.model.Shelter;
+import team.awesome.pet.repository.ShelterRepository;
 import team.awesome.pet.service.ShelterService;
 
 @RestController
 @RequestMapping("/shelters")
 public class ShelterController {
 
-  @Autowired
-  private ShelterService shelterService;
+	@Autowired
+//  private ShelterService shelterService;
+//  	private ShelterRepository shelterRepository;
+  	private ShelterService shelterService;
 
-  @RequestMapping(method = RequestMethod.GET)
-  public Collection<Shelter> getAllShelter() {
-    return this.shelterService.getAllShelter();
-  }
+	@RequestMapping(method = RequestMethod.GET)
+	public Collection<Shelter> getAllShelter() {
+		return this.shelterService.getAllShelter();
+	}
 
-  @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-  public Shelter getShelterById(@PathVariable("id") int id) {
-    return this.shelterService.getShelterById(id);
-  }
+	@RequestMapping(value = "/{id}", method = RequestMethod.GET)
+	public Shelter getShelterById(@PathVariable("id") int id) {
+		return this.shelterService.getShelterById(id);
+	}
 
-  @RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
-  public void insertShelter(@RequestBody Shelter shelter) {
-    this.shelterService.insertShelter(shelter);
-  }
+	@RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
+	public void insertShelter(@RequestBody Shelter shelter) {
+		this.shelterService.insertShelter(shelter);
+	}
 
-  @RequestMapping(method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE)
-  public void updateShelter(@RequestBody Shelter shelter) {
-    this.shelterService.updateShelter(shelter);
-  }
+	@RequestMapping(method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE)
+	public void updateShelter(@RequestBody Shelter shelter) {
+		this.shelterService.updateShelter(shelter);
+	}
 
-  @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-  public void deleteShelterById(@PathVariable("id") BigInteger id) {
-    this.shelterService.removeShelterById(id);
-  }
+	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
+	public void deleteShelterById(@PathVariable("id") int id) {
+		this.shelterService.removeShelterById(id);
+	}
 }
